Move TSnePoint to new package that will contain t-SNE reading stuff
